refactor: csdAssign & vsdVisibility moved to a upper folder since used in 2 services now
