+++ lib/keywords/csdAssign.js
